Moved class into new responses/instance_password.rb file
