+++ dcmgr/lib/dcmgr/endpoints/12.03/responses/instance.rb
@@ -123,20 +123,4 @@ module Dcmgr::Endpoints::V1203::Responses
       @instance.monitor_items
     end
   end
-
-  class InstancePassword < Dcmgr::Endpoints::ResponseGenerator
-    def initialize(instance)
-      raise ArgumentError, "Instance must be a #{Dcmgr::Models::Instance}" if !instance.is_a?(Dcmgr::Models::Instance)
-      @instance = instance
-    end
-
-    def generate
-      {
-        :id => @instance.canonical_uuid,
-        :encrypted_password => @instance.encrypted_password
-      }
-      # TODO: Clear the password field at specified instances record
-      #       if the flag delete_password_on_request flag is set.
-    end
-  end
 end
